Add listener.rs
