+++ src/server/state.rs
@@ -1,11 +1,7 @@
-use serde::Deserialize;
 use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::fs;
-use std::io::prelude::*;
-use std::os::unix::net::{UnixListener, UnixStream};
 use std::sync::mpsc::{channel, Sender};
-use std::thread;
 
 use super::worker;
 
@@ -14,13 +10,11 @@ use super::{
     task::Task,
     watcher::Watcher,
     worker::Action,
-    Communication,
     Message,
 };
 
 #[derive(Debug)]
 pub struct State {
-    pub listener: UnixListener,
     pub tasks: HashMap<String, ReadTask>,
     pub workers: HashMap<String, Sender<Action>>,
 }
@@ -32,30 +26,11 @@ impl Drop for State {
 }
 
 impl State {
-    pub fn new(socket: &str) -> Self {
+    pub fn new() -> Self {
         State {
-            listener: UnixListener::bind(socket).unwrap(),
             tasks: HashMap::new(),
             workers: HashMap::new(),
         }
-    }
-
-    pub fn listen(&mut self, sender: Sender<Communication>) {
-        let listener = self.listener.try_clone().unwrap();
-        thread::spawn(move || {
-            for stream in listener.incoming() {
-                match stream {
-                    Ok(stream) => {
-                        let s = sender.clone();
-                        thread::spawn(move || process_message(stream, s));
-                    }
-                    Err(err) => {
-                        println!("Error: {}", err);
-                        break;
-                    }
-                }
-            }
-        });
     }
 
     pub fn reload(&mut self, watcher: &Watcher) {
@@ -105,22 +80,4 @@ impl State {
             chan.send("\n----------\n".to_string()).unwrap();
         }
     }
-}
-
-fn process_message(stream: UnixStream, sender: Sender<Communication>) {
-    println!("Ready to recieve.");
-    let mut response = stream.try_clone().expect("Couldn't clone socket");
-    let mut de = serde_json::Deserializer::from_reader(stream);
-
-
-    if let Ok(msg) = Message::deserialize(&mut de) {
-        println!("Recieved {:?}", msg);
-        let (snd, receiver) = channel();
-        let com = Communication{ message: msg, channel: Some(snd)};
-        sender.send(com).unwrap();
-        for res in receiver.iter() {
-            response.write_all(res.as_bytes()).unwrap();
-        }
-    }
-    println!("End of transmission.");
 }
